+++ source/projects/chuck/core/chuck_compile.cpp
@@ -623,8 +623,11 @@ std::string Chuck_Compiler::resolveFilename( const std::string & filename,
     vector<string> exts; exts.push_back(".ck"); exts.push_back(".chug"); exts.push_back(".chug.wasm");
 
     // log
-    EM_log( CK_LOG_FINE, "resolving filename: '%s' importer: '%s' expand-search: %s",
-            filename.c_str(), importerAbsolutePath.c_str(), expandSearchToGlobal ? "YES" : "NO" );
+    EM_log( CK_LOG_FINE, "resolving filename: '%s'", filename.c_str() );
+    EM_pushlog();
+    EM_log( CK_LOG_FINE, "importer: '%s'", importerAbsolutePath.c_str() );
+    EM_log( CK_LOG_FINE, "expand-search: %s", expandSearchToGlobal ? "YES" : "NO" );
+    EM_poplog();
 
     // check if already absolute path
     if( isAlreadyAbsolutePath )
@@ -1639,9 +1642,9 @@ t_CKBOOL Chuck_Compiler::import_chugin_opt( const string & path, const string & 
     // get env
     Chuck_Env * env = this->env();
 
-    // NOTE this (verbose >= 5) is more informative if the chugin crashes, we can see the name
+    // NOTE this (running with verbose >= 5) is more informative if the chugin crashes while loading, we can see the name
     EM_log( CK_LOG_INFO, "@import loading [chugin] %s...", name.c_str() );
-    EM_log( CK_LOG_FINE, " |- path: '%s'", path.c_str() );
+    EM_log( CK_LOG_INFO, " |- path: '%s'", path.c_str() );
 
     // create chuck DLL data structure
     Chuck_DLL * dll = new Chuck_DLL( this->carrier(), name != "" ? name.c_str() : (extract_filepath_file(path)).c_str() );