+++ source/projects/chuck/core/chuck_type.cpp
@@ -3449,8 +3449,13 @@ t_CKTYPE type_engine_check_exp_unary( Chuck_Env * env, a_Exp_Unary unary )
         break;
 
         case ae_op_minus:
-            // float
-            if( isa( t, env->ckt_float ) ) return t;
+            // float, vec2
+            if( 
+                isa( t, env->ckt_float ) ||
+                isa( t, env->ckt_vec2 ) ||
+                isa( t, env->ckt_vec3 ) ||
+                isa( t, env->ckt_vec4 )
+            ) return t;
         case ae_op_tilda:
         case ae_op_exclamation:
             // int
@@ -5981,7 +5986,7 @@ Chuck_Namespace::~Chuck_Namespace()
 void Chuck_Namespace::add_type( const std::string & xid, Chuck_Type * theType )
 {
     // log it
-    EM_log( CK_LOG_DEBUG, "namespace '%s' adding type '%s'->'%s'", this->name.c_str(), xid.c_str(), theType->name().c_str() );
+    EM_log( CK_LOG_FINER, "namespace '%s' adding type '%s'->'%s'", this->name.c_str(), xid.c_str(), theType->name().c_str() );
     // add it
     this->type.add( xid, theType );
 }
@@ -5996,7 +6001,7 @@ void Chuck_Namespace::add_type( const std::string & xid, Chuck_Type * theType )
 void Chuck_Namespace::add_value( const std::string & xid, Chuck_Value * theValue )
 {
     // log it
-    EM_log( CK_LOG_DEBUG, "namespace '%s' adding value '%s'->'%s'", this->name.c_str(), xid.c_str(), theValue->name.c_str() );
+    EM_log( CK_LOG_FINER, "namespace '%s' adding value '%s'->'%s'", this->name.c_str(), xid.c_str(), theValue->name.c_str() );
     // add it
     this->value.add( xid, theValue );
 }
@@ -6011,7 +6016,7 @@ void Chuck_Namespace::add_value( const std::string & xid, Chuck_Value * theValue
 void Chuck_Namespace::add_func( const std::string & xid, Chuck_Func * theFunc )
 {
     // log it
-    EM_log( CK_LOG_DEBUG, "namespace '%s' adding func '%s'->'%s'", this->name.c_str(), xid.c_str(), theFunc->base_name.c_str() );
+    EM_log( CK_LOG_FINER, "namespace '%s' adding func '%s'->'%s'", this->name.c_str(), xid.c_str(), theFunc->base_name.c_str() );
     // add it
     this->func.add( xid, theFunc );
 }
@@ -8188,6 +8193,9 @@ void type_engine_init_op_overload_builtin( Chuck_Env * env )
     registry->reserve( NULL, ae_op_minusminus, env->ckt_float );
     registry->reserve( NULL, ae_op_minus, env->ckt_int );
     registry->reserve( NULL, ae_op_minus, env->ckt_float );
+    registry->reserve( NULL, ae_op_minus, env->ckt_vec2 );
+    registry->reserve( NULL, ae_op_minus, env->ckt_vec3 );
+    registry->reserve( NULL, ae_op_minus, env->ckt_vec4 );
     registry->reserve( NULL, ae_op_tilda, env->ckt_int );
     registry->reserve( NULL, ae_op_exclamation, env->ckt_int );
     registry->reserve( NULL, ae_op_new, env->ckt_object );